File controller y datos de peticion api en un txt

+++ src/main/java/com/capitalmarkets/app/core/services/CurrencyControllerService.java
@@ -23,7 +23,7 @@ public class CurrencyControllerService implements IcurrencyControllerService{
     }
 
     @Override
-    public CurrencyConverterDTO getConversion(int value, String base, String conversion) {
+    public CurrencyConverterDTO getConversion(double value, String base, String conversion) {
         return servicio.getConversion(value,base,conversion);
     }
