+++ src/main/java/com/capitalmarkets/app/core/services/IcurrencyControllerService.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface IcurrencyControllerService {
 
     List<CurrencyApiDTO> getAll();
-    CurrencyConverterDTO getConversion(int value, String base, String conversion);
+    CurrencyConverterDTO getConversion(double value, String base, String conversion);
     CurrencyHistoricalDTO getHistorical(String date, double value, String base, String conversion);
 }
